Fix some minor infelicities
